refactor: move middlewares to capture/middlewares
Move {cleaner,header,pcap}_middlewares into src/capture/middlewares
folder.
