+++ src/capture/middlewares/cleaner_middleware/CMakeLists.txt
