+++ src/capture/middlewares/header_middleware/dns_decoder.c
@@ -36,10 +36,10 @@
 #include <netinet/udp.h>
 #include <arpa/inet.h>
 
-#include "../../utils/utarray.h"
-#include "../../utils/allocs.h"
-#include "../../utils/os.h"
-#include "../../utils/hash.h"
+#include "../../../utils/utarray.h"
+#include "../../../utils/allocs.h"
+#include "../../../utils/os.h"
+#include "../../../utils/hash.h"
 
 #include "packet_decoder.h"
