+++ src/capture/middlewares/header_middleware/header_middleware.h
@@ -18,21 +18,19 @@
  ****************************************************************************/
 
 /**
- * @file cleaner_middleware.h
+ * @file header_middleware.h
  * @author Alexandru Mereacre
- * @brief File containing the definition of the middleware cleaner utilities.
+ * @brief File containing the definition of the header middleware utilities.
  */
 
-#ifndef CLEANER_MIDDLEWARE_H
-#define CLEANER_MIDDLEWARE_H
+#ifndef HEADER_MIDDLEWARE_H
+#define HEADER_MIDDLEWARE_H
 
-#include "../middleware.h"
+#include "../../middleware.h"
 
 /**
- * @brief Cleaner Middleware.
+ * @brief Header Middleware.
  * @author Alexandru Mereacre, Alois Klink
- * The cleaner middleware is designed to periodically remove the oldest
- * PCAP files when the use more than `CLEANER_STORE_SIZE` KiB.
  */
-extern struct capture_middleware cleaner_middleware;
+extern struct capture_middleware header_middleware;
 #endif
