+++ src/capture/middlewares/header_middleware/mdns_decoder.h
@@ -26,9 +26,9 @@
 #ifndef MDNS_DECODER_H
 #define MDNS_DECODER_H
 
-#include "../../utils/net.h"
-#include "../../utils/os.h"
-#include "../../utils/utarray.h"
+#include "../../../utils/net.h"
+#include "../../../utils/os.h"
+#include "../../../utils/utarray.h"
 
 #include "packet_decoder.h"
