+++ src/capture/middlewares/header_middleware/packet_decoder.h
@@ -41,10 +41,10 @@
 #include <arpa/inet.h>
 #include <pcap.h>
 
-#include "../../utils/utarray.h"
-#include "../../utils/allocs.h"
-#include "../../utils/os.h"
-#include "../../utils/net.h"
+#include "../../../utils/utarray.h"
+#include "../../../utils/allocs.h"
+#include "../../../utils/os.h"
+#include "../../../utils/net.h"
 
 #define MAX_QUESTION_LEN 255
