+++ src/capture/middlewares/header_middleware/packet_queue.h
@@ -31,7 +31,7 @@
 #include <string.h>
 
 #include "packet_decoder.h"
-#include "../../utils/list.h"
+#include "../../../utils/list.h"
 
 /**
  * @brief Packet queueu structure definition
