+++ src/capture/middlewares/pcap_middleware/pcap_middleware.c
@@ -35,13 +35,13 @@
 #include "pcap_queue.h"
 #include "sqlite_pcap.h"
 
-#include "../../utils/allocs.h"
-#include "../../utils/os.h"
-#include "../../utils/log.h"
-#include "../../utils/squeue.h"
-#include "../../utils/eloop.h"
+#include "../../../utils/allocs.h"
+#include "../../../utils/os.h"
+#include "../../../utils/log.h"
+#include "../../../utils/squeue.h"
+#include "../../../utils/eloop.h"
 
-#include "../pcap_service.h"
+#include "../../pcap_service.h"
 
 #define PCAP_SUBFOLDER_NAME                                                    \
   "./capture" /* Subfodler name to store raw pcap data                         \
