+++ src/capture/middlewares/pcap_middleware/pcap_queue.h
@@ -32,7 +32,7 @@
 #include <pcap.h>
 #include <stdint.h>
 
-#include "../../utils/list.h"
+#include "../../../utils/list.h"
 
 /**
  * @brief pcap queueu structure definition
