+++ src/capture/middlewares/pcap_middleware/sqlite_pcap.c
@@ -31,10 +31,10 @@
 
 #include "sqlite_pcap.h"
 
-#include "../../utils/allocs.h"
-#include "../../utils/os.h"
-#include "../../utils/log.h"
-#include "../../utils/sqliteu.h"
+#include "../../../utils/allocs.h"
+#include "../../../utils/os.h"
+#include "../../../utils/log.h"
+#include "../../../utils/sqliteu.h"
 
 int init_sqlite_pcap_db(sqlite3 *db) {
   int rc;
