+++ tests/capture/middlewares/test_header_middleware.c
@@ -14,8 +14,8 @@
 
 #include "utils/log.h"
 #include "utils/sqliteu.h"
-#include "capture/header_middleware/header_middleware.h"
-#include "capture/header_middleware/sqlite_header.h"
+#include "capture/middlewares/header_middleware/header_middleware.h"
+#include "capture/middlewares/header_middleware/sqlite_header.h"
 
 extern int __real_sqlite3_open(const char *filename, sqlite3 **ppDb);
