+++ tests/capture/middlewares/test_packet_queue.c
@@ -13,7 +13,7 @@
 #include <cmocka.h>
 
 #include "utils/log.h"
-#include "capture/header_middleware/packet_queue.h"
+#include "capture/middlewares/header_middleware/packet_queue.h"
 
 static void test_push_packet_queue(void **state) {
   (void)state; /* unused */
