+++ tests/capture/middlewares/test_sqlite_pcap.c
@@ -14,7 +14,7 @@
 
 #include "utils/log.h"
 #include "utils/sqliteu.h"
-#include "capture/pcap_middleware/sqlite_pcap.h"
+#include "capture/middlewares/pcap_middleware/sqlite_pcap.h"
 
 static void test_open_sqlite_pcap_db(void **state) {
   (void)state; /* unused */
